Fix example (int --> float)

+++ 2024-11-11_0001/main-3.cpp
@@ -32,7 +32,7 @@ public:
         return Square(sqrt(resultArea));
     }
 
-    Square operator+(int otherArea) const
+    Square operator+(float otherArea) const
     {
         float resultArea = GetArea() + otherArea;
